requirements_raw.txt
qrcode
pypng
python-telegram-bot
ecdsa
nest_asyncio
pyserial
websocket-client
rel